merge: 리펙토링 : Unit test 작성 (#31)
* task : joinCustomer

* task : getCustomerById

* task : getCustomerByIdNotFound -r

* task : getAllCustomers

* task : updateCustomer

* task : countCustomers

* task : getAllItems

* task : getAllItemsSortedByPrice

* task : getAllItemsSortedByName

* task : getItemById

* task : getItemsById_NotFound

* task : addItem

* task : countItems

* feature : OrderControllerTest / refactor : entity -> item, order

* refactor : rename

---------

+++ backend/project1/src/main/java/com/team4/project1/domain/item/controller/ApiV1ItemController.java
@@ -16,7 +16,7 @@ public class ApiV1ItemController {
     private final ItemService itemService;
 
     @GetMapping("/{itemId}")
-    public ItemDto item(@PathVariable Integer itemId) {
+    public ItemDto item(@PathVariable Long itemId) {
         return itemService.getItemById(itemId).orElseThrow(() -> new ItemNotFoundException(itemId));
     }
