+++ backend/project1/src/main/java/com/team4/project1/global/exception/ItemNotFoundException.java
@@ -1,7 +1,7 @@
 package com.team4.project1.global.exception;
 
 public class ItemNotFoundException extends RuntimeException {
-    public ItemNotFoundException(Integer id) {
+    public ItemNotFoundException(Long id) {
         super("해당 상품을 찾을 수 없습니다. (ID: " + id + ")");
     }
 }
